refactor: tighten agent harness surfaces
